Update 3_Training_Neural_Networks.ipynb (#295)
Your equation is off for the gradients of z with respect to x. It just so happens that with the given number of elements in the chosen x matrix you end up with the right result. In general however it is not the case. 

These changes updates the equation to give the right gradient, and also displays the two matrices: x.grad and 2*x/torch.numel(x) side by side for the student to be able to play around with the input x and see for themselves that this is the right gradient.

Co-authored-by: Nicki Skafte Detlefsen <skaftenicki@gmail.com>